chore(disctl): rename daisctl to disctl to match with the rest of the platform components names
