+++ .github/workflows/disctl-release.yml
@@ -1,12 +1,12 @@
-name: Release Daisctl
+name: Release Disctl
 
 env:
-  WORKDIR: tools/daisctl
+  WORKDIR: tools/disctl
 
 on:
   push:
     tags:
-      - "*-daisctl"
+      - "*-disctl"
 
 permissions:
   contents: write
@@ -22,6 +22,7 @@ jobs:
       - uses: actions/setup-go@d35c59abb061a4a6fb18e82ac0862c26744d6ab5 # v5.5.0
         with:
           go-version-file: '${{ env.WORKDIR }}/go.mod'
+          cache-dependency-path: '${{ env.WORKDIR }}/go.sum'
           cache: true
       - uses: goreleaser/goreleaser-action@e435ccd777264be153ace6237001ef4d979d3a7a # v6.4.0
         with:
