+++ tools/disctl/pkg/altinn/environments_test.go
