+++ tools/disctl/pkg/altinn/testdata/testgetenvironments/malformed_json_response.json
