+++ tools/disctl/pkg/altinn/testdata/testgetenvironments/server_error.json
