+++ tools/disctl/pkg/altinn/testdata/testgetenvironments/valid_response.json
