+++ tools/disctl/pkg/altinn/testdata/testinitappsdata/invalid_response_from_kubewrapper_result_in_error.json
