+++ tools/disctl/pkg/altinn/testdata/testinitappsdata/no_deployments_returned_from_kubewrapper.json
