+++ tools/disctl/pkg/altinn/testdata/testinitappsdata/server_error_from_kubewrapper_result_in_error.json
