+++ tools/disctl/pkg/altinn/testdata/testinitappsdata/valid_response_from_kubewrapper.json
