+++ tools/disctl/pkg/util/constants.go
